requirements.txt
kafka-python==2.0.2
pyspark==3.2.0
apache-flink==1.14.4
pandas==1.4.2
sqlalchemy==1.4.36
pymongo==4.1.1
apache-airflow==2.2.5
requests==2.27.1
textblob==0.15.3
prometheus-client==0.14.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3